+++ addons/shared_addons/modules/feedback_manager/controllers/feedback_manager.php
@@ -155,20 +155,15 @@ class Feedback_manager extends Public_Controller {
             $deleted_ids = array();
             foreach ($ids as $id){
                 if ($post = $this->feedback_manager_m->get($id)){
-                    if ($this->feedback_manager_m->delete($id)){
-                        $postx = $this->feedbackuser_m->get_many_by(array('feedback_manager_id'=>$id));
-                        foreach ($postx as $keyx) {
-                            $this->feedbackuser_m->delete($postx->id);
+                    $postx = $this->feedbackuser_m->get_by(array('feedback_manager_id'=>$id));
+                    $posty = $this->feedback_manager_question_m->get_by(array('feedback_manager_id'=>$id));
+                    if($postx == "" && $posty == "")
+                    {
+                        if ($this->feedback_manager_m->delete($id)){
+                            $this->pyrocache->delete('feedback_manager_m');
+                            $post_names[] = $post->title;
+                            $deleted_ids[] = $id;
                         }
-                        $posty = $this->feedback_manager_question_m->get_many_by(array('feedback_manager_id'=>$id));
-                        foreach ($posty as $key) {
-                            $this->feedback_manager_question_m->delete($key->id);
-                        }
-                        $this->pyrocache->delete('feedback_manager_question_m');
-                        $this->pyrocache->delete('feedbackuser_m');
-                        $this->pyrocache->delete('feedback_manager_m');
-                        $post_names[] = $post->title;
-                        $deleted_ids[] = $id;
                     }
                 }
             }
